latest: use an absolute redirect, update tests
This avoids an issue where the redirect is interpreted as a relative
URL.
This is caused by an update in one of the dependencies.
Resolves #28.

While I'm at it, updated one of the tests, because recent versions of
Werkzeug return a 308 for a permanent redirect. This is like a 301, but
preserves the request, including if it is a POST.